Refactor: (상대방) 숫자를 생성하는 기능 분리
feat:
- 상대방이 숫자를 랜덤으로 생성하는 로직을 Computer.js로 분리하여 구현하였습니다.
- 최종적으로 Computer 클래스를 생성해, 조건에 맞는 3자리의 숫자를 랜덤으로 생성하는 기능을 구현하였습니다.
